requirements.txt
# MarketFlow — Full Stack Dependencies
# Python 3.11+

# Flask 서버
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
bcrypt>=4.0.0

# 데이터 수집
pykrx>=1.0.45
requests>=2.31.0
beautifulsoup4>=4.12.0
finance-datareader>=0.9.50
finnhub-python>=2.4.0
fredapi>=0.5.0

# 데이터 처리
pandas>=2.0.0
numpy>=2.1.0
yfinance
scipy>=1.11.0
scikit-learn>=1.3.0

# 시각화 (portfolio_risk, backtest)
matplotlib>=3.7.0
seaborn>=0.13.0

# 스케줄링
schedule>=1.2.0

# AI/LLM
google-generativeai>=0.3.0
openai>=1.0.0

# 기타
python-dotenv>=1.0.0
tqdm>=4.66.0
openpyxl>=3.1.0

# 결제
stripe>=7.0.0

# 배포/운영
gunicorn>=21.2.0
aiohttp>=3.9.0
setuptools
certifi>=2024.0.0
filelock>=3.0.0
